+++ modules/storage/main.tf
@@ -14,10 +14,14 @@ resource "azurerm_storage_account" "this" {
   account_replication_type = var.account_replication_type
 
   public_network_access_enabled = var.public_network_access_enabled
-  network_rules {
-    default_action             = var.network_rules_default_action
-    ip_rules                   = var.public_ip_allow_list
-    virtual_network_subnet_ids = var.virtual_network_subnet_ids
+  dynamic "network_rules" {
+    for_each = var.network_rules_default_action == "Deny" || length(var.public_ip_allow_list) > 0 || var.virtual_network_subnet_ids != null ? ["rule"] : []
+
+    content {
+      default_action             = var.network_rules_default_action
+      ip_rules                   = var.public_ip_allow_list
+      virtual_network_subnet_ids = var.virtual_network_subnet_ids
+    }
   }
 
   tags = var.tags